requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.18.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.0
asttokens==3.0.1
astunparse==1.6.3
asyncer==0.0.8
attrs==25.4.0
autograd==1.8.0
babel==2.17.0
black==26.1.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cuda-bindings==12.9.4; sys_platform == 'linux'
cuda-pathfinder==1.3.3; sys_platform == 'linux'
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
deprecation==2.1.0
diskcache==5.6.3
distro==1.9.0
docutils==0.22.4
dspy==3.1.2
dspy-ai==3.1.2
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.135.1
fastjsonschema==2.21.2
fastuuid==0.14.0
feedparser==6.0.12
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
gepa==0.0.24
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.60.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
hnswlib==0.8.0; sys_platform == 'linux'
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.3.4
humanfriendly==10.0
idna==3.11
imagesize==1.4.1
importlib_metadata==8.7.1
inferactively-pymdp==0.0.7.1
iniconfig==2.3.0
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json_repair==0.55.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-cache==1.0.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
lancedb==0.27.1
litellm==1.81.6
lmql==0.7.3
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.26.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
myst-nb==1.3.0
myst-parser==5.0.0
nbclient==0.10.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
nose==1.3.7
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1; sys_platform == 'linux'
nvidia-cuda-cupti-cu12==12.8.90; sys_platform == 'linux'
nvidia-cuda-nvrtc-cu12==12.8.93; sys_platform == 'linux'
nvidia-cuda-runtime-cu12==12.8.90; sys_platform == 'linux'
nvidia-cudnn-cu12==9.10.2.21; sys_platform == 'linux'
nvidia-cufft-cu12==11.3.3.83; sys_platform == 'linux'
nvidia-cufile-cu12==1.13.1.3; sys_platform == 'linux'
nvidia-curand-cu12==10.3.9.90; sys_platform == 'linux'
nvidia-cusolver-cu12==11.7.3.90; sys_platform == 'linux'
nvidia-cusparse-cu12==12.5.8.93; sys_platform == 'linux'
nvidia-cusparselt-cu12==0.7.1; sys_platform == 'linux'
nvidia-nccl-cu12==2.27.5; sys_platform == 'linux'
nvidia-nvjitlink-cu12==12.8.93; sys_platform == 'linux'
nvidia-nvshmem-cu12==3.4.5; sys_platform == 'linux'
nvidia-nvtx-cu12==12.8.90; sys_platform == 'linux'
onnxruntime==1.23.2
openai==2.16.0
openpyxl==3.1.5
optuna==4.7.0
orjson==3.11.6
overrides==7.7.0
packaging==26.0
pandas==3.0.0
parso==0.8.5
pathspec==1.0.4
pexpect==4.9.0; sys_platform == 'linux'
pillow==12.1.0
platformdirs==4.5.1
playwright==1.58.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.2.1
ptyprocess==0.7.0; sys_platform == 'linux'
pure_eval==0.2.3
py==1.11.0
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pymdp==0.0.1
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytokens==0.4.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
roman-numerals==4.1.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
sentence-transformers==5.2.2
setuptools==82.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==9.0.4
sphinx_rtd_theme==3.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.46
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
triton==3.6.0; sys_platform == 'linux'
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.5.0
websockets==15.0.1
wheel==0.46.3
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
markitdown[all]>=0.1.5
